LogoTool ES class
